Creates ShipType and Ship
With this approach, it is necessary to be sure that an instance of a Ship is
the same in all necessary positions in the matrix, this way when that instance
is Hit we can calculate when it sinks.

We can guarantee this when populating the Board since both DrawHorizontal and DrawVertical
commands can populate each position with the same instance of a ship.