Populates board with ships and verifies colision between ships
This is a naive approach to verify colision between ships, when a
colision is confirmed it recalculates a new place to populate the ship
on the board.

This approach is not ideal when there is high demand to create new
boards, it takes an average of 6 cycles to produce one board.

A better approach may be to verify which positions have been populated
and discard them.